Rename package, move directories
