+++ app/build.gradle
@@ -1,7 +1,8 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'kotlin-android-extensions'
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'kotlin-kapt'
+}
 
 
 android {
@@ -13,7 +14,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.example.rickandmortycharacters"
+        applicationId "com.ntz.rickandmortycharacters"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
@@ -51,15 +52,15 @@ dependencies {
     implementation("com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:0.9.2")
     implementation "com.squareup.retrofit2:converter-gson:$gson_version"
 
-    api 'com.google.dagger:dagger:2.30.1'
-    api 'com.google.dagger:dagger-android:2.30.1'
-    api 'com.google.dagger:dagger-android-support:2.30.1'
+    api "com.google.dagger:dagger:$dagger_version"
+    api "com.google.dagger:dagger-android:$dagger_version"
+    api "com.google.dagger:dagger-android-support:$dagger_version"
 
-    annotationProcessor 'com.google.dagger:dagger-compiler:2.30.1'
-    kapt 'com.google.dagger:dagger-compiler:2.30.1'
+    annotationProcessor "com.google.dagger:dagger-compiler:$dagger_version"
+    kapt "com.google.dagger:dagger-compiler:$dagger_version"
 
-    annotationProcessor 'com.google.dagger:dagger-android-processor:2.30.1'
-    kapt 'com.google.dagger:dagger-android-processor:2.30.1'
+    annotationProcessor "com.google.dagger:dagger-android-processor:$dagger_version"
+    kapt "com.google.dagger:dagger-android-processor:$dagger_version"
 
     compileOnly 'javax.annotation:jsr250-api:1.0'
     implementation 'javax.inject:javax.inject:1'
