+++ app/src/main/java/com/ntz/rickandmortycharacters/data/network/model/ApiListResult.kt
@@ -1,0 +1,5 @@
+package com.ntz.rickandmortycharacters.data.network.model
+
+class ApiListResult<T> {
+    var results = emptyList<T>()
+}
