+++ app/src/main/java/com/ntz/rickandmortycharacters/data/network/model/ResultModel.kt
@@ -1,0 +1,3 @@
+package com.ntz.rickandmortycharacters.data.network.model
+
+data class ResultModel(val name: String, val image: String)
