+++ app/src/main/java/com/ntz/rickandmortycharacters/data/network/services/CharacterApi.kt
@@ -1,0 +1,11 @@
+package com.ntz.rickandmortycharacters.data.network.services
+
+import com.ntz.rickandmortycharacters.data.network.model.ApiListResult
+import com.ntz.rickandmortycharacters.data.network.model.ResultModel
+import retrofit2.http.GET
+
+
+interface CharacterApi {
+    @GET("character")
+    suspend fun getCharacters(): ApiListResult<ResultModel>
+}
