+++ app/src/main/java/com/ntz/rickandmortycharacters/data/network/services/CharacterRepository.kt
@@ -1,0 +1,12 @@
+package com.ntz.rickandmortycharacters.data.network.services
+
+import com.ntz.rickandmortycharacters.data.network.model.ApiListResult
+import com.ntz.rickandmortycharacters.data.network.model.ResultModel
+import javax.inject.Inject
+
+
+class CharacterRepository @Inject constructor(private val api: CharacterApi) {
+    suspend fun getCharacters() : ApiListResult<ResultModel> {
+        return api.getCharacters()
+    }
+}
