+++ app/src/main/java/com/ntz/rickandmortycharacters/di/character/CharacterListComponent.kt
@@ -1,0 +1,11 @@
+package com.ntz.rickandmortycharacters.di.character
+
+import com.ntz.rickandmortycharacters.ui.list.CharListViewModel
+import dagger.Subcomponent
+
+
+@Subcomponent(modules = [CharacterModule::class])
+@CharacterScope
+interface CharacterListComponent {
+    fun inject(viewModel: CharListViewModel)
+}
