+++ app/src/main/java/com/ntz/rickandmortycharacters/di/character/CharacterScope.kt
@@ -1,0 +1,7 @@
+package com.ntz.rickandmortycharacters.di.character
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class CharacterScope
