+++ app/src/main/java/com/ntz/rickandmortycharacters/di/common/AppModule.kt
@@ -1,0 +1,14 @@
+package com.ntz.rickandmortycharacters.di.common
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class AppModule(private val context: Context) {
+
+    @Provides
+    @Singleton
+    fun provideContext() = context
+}
