+++ app/src/main/java/com/ntz/rickandmortycharacters/di/common/NetworkModule.kt
@@ -1,0 +1,25 @@
+package com.ntz.rickandmortycharacters.di.common
+
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import dagger.Module
+import dagger.Provides
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import com.ntz.rickandmortycharacters.utils.Constants
+import javax.inject.Singleton
+
+@Module
+class NetworkModule {
+    @Provides
+    @Singleton
+    fun provideRetrofit(): Retrofit =
+        buildClient()
+
+    private fun buildClient(): Retrofit {
+        return Retrofit.Builder()
+            .addConverterFactory(GsonConverterFactory.create())
+            .addCallAdapterFactory(CoroutineCallAdapterFactory())
+            .baseUrl(Constants.BASE_URL)
+            .build()
+    }
+}
