+++ app/src/main/java/com/ntz/rickandmortycharacters/utils/Constants.kt
@@ -1,0 +1,5 @@
+package com.ntz.rickandmortycharacters.utils
+
+object Constants {
+    const val BASE_URL = "https://rickandmortyapi.com/api/"
+}
