--- app/src/main/java/data/network/model/ApiListResult.kt
@@ -1,5 +1,0 @@
-package data.network.model
-
-class ApiListResult<T> {
-    var results = emptyList<T>()
-}
