--- app/src/main/java/data/network/model/ResultModel.kt
@@ -1,3 +1,0 @@
-package data.network.model
-
-data class ResultModel(val name: String, val image: String)
