--- app/src/main/java/di/common/AppModule.kt
@@ -1,14 +1,0 @@
-package di.common
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class AppModule(private val context: Context) {
-
-    @Provides
-    @Singleton
-    fun provideContext() = context
-}
