--- app/src/main/java/di/common/NetworkModule.kt
@@ -1,26 +1,0 @@
-package di.common
-
-import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-import dagger.Component
-import dagger.Module
-import dagger.Provides
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import utils.Constants
-import javax.inject.Singleton
-
-@Module
-class NetworkModule {
-    @Provides
-    @Singleton
-    fun provideRetrofit(): Retrofit =
-        buildClient()
-
-    private fun buildClient(): Retrofit {
-        return Retrofit.Builder()
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
-            .baseUrl(Constants.BASE_URL)
-            .build()
-    }
-}
