--- app/src/main/java/di/common/character/CharacterListComponent.kt
@@ -1,12 +1,0 @@
-package di.common.character
-
-import com.example.rickandmortycharacters.ui.list.CharListViewModel
-import dagger.Subcomponent
-import javax.inject.Singleton
-
-
-@Subcomponent(modules = [CharacterModule::class])
-@CharacterScope
-interface CharacterListComponent {
-    fun inject(viewModel: CharListViewModel)
-}
