--- app/src/main/java/di/common/character/CharacterScope.kt
@@ -1,7 +1,0 @@
-package di.common.character
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class CharacterScope
