--- app/src/main/java/utils/Constants.kt
@@ -1,5 +1,0 @@
-package utils
-
-object Constants {
-    const val BASE_URL = "https://rickandmortyapi.com/api/"
-}
